Fix URLService
The web service we call returns a set of URLs that will then return a
SMIL with either a video on Akamai or on video.foodnetwork.ca. The
former video URL returns a 403, the latter will play. Modified the code
to only use the latter. Plus some debugging info and comments.